Files
All the folder's and view's are better organized then before apart from scriptorium that needs a refresh in another branch
