+++ Portfolio/src/views/about/aboutMe/aboutMe.css
@@ -1,0 +1,66 @@
+/*  */
+/* Canvas html stuff */
+/*  */
+
+.play-ground {
+  position: absolute;
+  cursor: default;
+  width: 100%;
+  height: 100%;
+  z-index: 0;
+}
+
+.shadow {
+  position: absolute;
+  cursor: default;
+  width: 100%;
+  height: 100%;
+  z-index: -1;
+}
+/*  */
+/* The about page animation is removed for mobile so this is what replaces it */
+/*  */
+.about-content {
+  position: relative;
+  width: 100%;
+  height: 225vh;
+  display: flex;
+  justify-content: flex-start;
+  flex-direction: column;
+  align-items: center;
+  overflow: hidden;
+  gap: var(--space-3);
+}
+
+.mobile-intro {
+  height: auto;
+  position: relative;
+  width: 90%;
+  border-radius: 25px;
+  padding: var(--space-2);
+  background-color: var(--color-4);
+}
+
+.mobile-intro.hidden {
+  display: none;
+}
+
+.mobile-intro h4 {
+  color: var(--color-2);
+  font-size: 2.5rem;
+  margin: var(--space-2) 0;
+}
+
+.mobile-intro p {
+  margin: var(--space-2) 0;
+  font-size: 1.25rem;
+  font-family: var(--font-body), sans-serif;
+  line-height: 1.3;
+  color: var(--font-middle);
+}
+
+@media (max-width: 1000px) {
+  .about-content {
+    height: auto;
+  }
+}
