+++ Portfolio/src/views/about/aboutMe/aboutMe.jsx
@@ -1,19 +1,22 @@
 "use client";
 import PlayGround from "./playground/playGround";
 import { useState, useEffect } from "react";
-import PropTypes from "prop-types";
-import "./about.css";
+import "./aboutMe.css";
 
-export default function AboutMobile({
+import {
   Header1,
   Header2,
   Header3,
   MainText1,
   MainText2,
   MainText3,
-}) {
-  const [ToSmall, setToSmall] = useState(false);
+} from "../text";
 
+function AboutMe() {
+  //
+  // Mobile alternative
+  //
+  const [ToSmall, setToSmall] = useState(false);
   useEffect(() => {
     function handleResize() {
       setToSmall(window.innerWidth < 1000);
@@ -26,6 +29,9 @@ export default function AboutMobile({
   return (
     <div className="about-content">
       {ToSmall === false && <PlayGround />}
+      {/*  */}
+      {/* Mobile alternative */}
+      {/*  */}
       <div className={`mobile-intro${ToSmall ? "" : " hidden"}`}>
         <h4>{Header1}</h4>
         <p>{MainText1}</p>
@@ -42,11 +48,4 @@ export default function AboutMobile({
   );
 }
 
-AboutMobile.propTypes = {
-  Header1: PropTypes.string.isRequired,
-  Header2: PropTypes.string.isRequired,
-  Header3: PropTypes.string.isRequired,
-  MainText1: PropTypes.string.isRequired,
-  MainText2: PropTypes.string.isRequired,
-  MainText3: PropTypes.string.isRequired,
-};
+export default AboutMe;
