+++ Portfolio/src/views/contact/contact.css
@@ -1,18 +1,7 @@
 /*  */
-/* Background canvas */
-/*  */
-#contact-canvas {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  z-index: 5;
-}
-/*  */
 /* Main container */
 /*  */
-.contact {
+.contact-container {
   display: flex;
   align-items: center;
   height: auto;
@@ -20,266 +9,4 @@
   margin: 7.5rem 0;
   flex-direction: column;
   gap: var(--space-7);
-}
-/*  */
-/* Intro section */
-/*  */
-.contact-intro {
-  display: flex;
-  width: 90%;
-  margin: 2.5rem 0;
-  margin-bottom: 7rem;
-}
-/*  */
-/* Left side text and email */
-.contact-intro-left {
-  display: flex;
-  flex-direction: column;
-  width: 60%;
-}
-
-.contact-intro-left h2 {
-  font-size: 12rem;
-  font-weight: 500;
-  color: var(--color-1);
-  line-height: 0.85;
-  margin: 0;
-}
-
-.contact-intro-left h2:nth-last-of-type(1) {
-  margin-left: 13rem;
-}
-
-.contact-intro-left-email {
-  width: auto;
-  height: auto;
-  margin: 0 auto;
-  margin-top: var(--space-4);
-  position: relative;
-  overflow: hidden;
-  cursor: pointer;
-}
-
-.contact-intro-left-email::after {
-  content: "";
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  width: 100%;
-  height: 2px;
-  background-color: var(--color-2);
-  transition:
-    height 0.25s var(--power-1-out),
-    background-color 0.25s var(--power-1-out);
-  z-index: -1;
-}
-
-.contact-intro-left-email:hover::after {
-  height: 100%;
-}
-
-.contact-intro-left-email h3 {
-  font-size: 1.5rem;
-  margin: 0;
-  margin-bottom: var(--space-0);
-  font-weight: 400;
-  color: var(--color-1);
-}
-
-/*  */
-/* Right side text */
-.contact-intro-right {
-  display: flex;
-  flex-direction: column;
-  width: 40%;
-}
-
-.contact-intro-right h3 {
-  font-size: 2rem;
-  margin-top: var(--space-5);
-  margin-bottom: var(--space-2);
-  margin-left: var(--space-2);
-  font-weight: 600;
-  color: var(--color-2);
-}
-
-.contact-intro-right-text {
-  padding: var(--space-2);
-  display: flex;
-  border-radius: 30px;
-  background-color: var(--color-4);
-}
-
-.contact-intro-right-text p {
-  font-size: 1rem;
-  margin: 0;
-  color: var(--font-middle);
-}
-
-/*  */
-/* Faq section */
-/*  */
-
-.contact-faq {
-  width: 90%;
-  height: auto;
-  display: flex;
-  flex-direction: column;
-}
-/*  */
-/* Faq container */
-.contact-faq-container {
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-1);
-  width: 100%;
-  min-height: 75vh;
-}
-/*  */
-/* Faq container item */
-.contact-faq-container-item {
-  display: flex;
-  background-color: var(--color-4);
-  padding: var(--space-3);
-  max-height: 200px;
-  overflow: hidden;
-  align-items: center;
-  border-radius: 25px;
-  transition:
-    background-color 0.5s var(--power-1-out),
-    max-height 0.5s var(--power-1-out);
-}
-
-.contact-faq-container-item h4 {
-  font-size: 1.5rem;
-  margin: 0;
-  color: var(--color-2);
-  margin-left: auto;
-  transition: font-size 0.5s var(--power-1-out);
-}
-
-.contact-faq-container-item-text {
-  width: 50%;
-  height: auto;
-}
-
-.contact-faq-container-item-text h3 {
-  font-size: 1.5rem;
-  margin: 0;
-  color: var(--color-1);
-  transition:
-    margin-bottom 0.5s var(--power-1-out),
-    color 0.5s var(--power-1-out);
-}
-
-.contact-faq-container-item-text p {
-  font-size: 1rem;
-  color: var(--font-middle);
-  max-height: 0;
-  overflow: hidden;
-  margin: 0;
-  transition: max-height 0.5s var(--power-1-out);
-}
-/* Hover state */
-.contact-faq-container-item.active {
-  background-color: var(--color-1);
-  max-height: 600px;
-}
-.contact-faq-container-item.active h4 {
-  font-size: 3rem;
-  color: var(--color-2);
-}
-
-.contact-faq-container-item-text.active h3 {
-  margin-bottom: var(--space-2);
-  color: var(--color-4);
-}
-
-.contact-faq-container-item-text.active p {
-  color: var(--color-3);
-  max-height: 300px;
-}
-
-/*  */
-/* Header section mobile changes */
-/*  */
-@media (max-width: 1700px) {
-  /* 10% decrease */
-  .contact-intro-left h2 {
-    font-size: 10.8rem;
-  }
-}
-
-@media (max-width: 1600px) {
-  /* 20% decrease */
-  .contact-intro-left h2 {
-    font-size: 9.6rem;
-  }
-}
-
-@media (max-width: 1500px) {
-  /* 30% decrease */
-  .contact-intro-left h2 {
-    font-size: 8.4rem;
-  }
-}
-
-@media (max-width: 1400px) {
-  .contact-intro {
-    flex-direction: column;
-  }
-  .contact-intro-left {
-    width: 75%;
-    margin: 0 auto;
-  }
-  .contact-intro-right {
-    width: 75%;
-    margin: 0 auto;
-  }
-
-  .contact-intro-left h2 {
-    margin: 0;
-    line-height: 1;
-  }
-
-  .contact-intro-left h2:nth-last-of-type(1) {
-    margin: 0;
-  }
-}
-
-@media (max-width: 700px) {
-  .contact-intro-left {
-    width: 100%;
-  }
-  .contact-intro-right {
-    width: 100%;
-    margin: 0 auto;
-  }
-}
-
-@media (max-width: 500px) {
-  /* 40% decrease */
-  .contact-intro-left h2 {
-    font-size: 7.2rem;
-  }
-  /* 10% decrease */
-  .contact-intro-right h3 {
-    font-size: 1.8rem;
-  }
-}
-
-@media (max-width: 400px) {
-  /* 50% decrease */
-  .contact-intro-left h2 {
-    font-size: 6rem;
-  }
-}
-/*  */
-/* Faq mobile changes */
-/*  */
-
-@media (max-width: 1200px) {
-  .contact-faq-container-item-text {
-    width: 70%;
-  }
 }
