+++ Portfolio/src/views/contact/faq/faq.css
@@ -1,0 +1,89 @@
+/*  */
+/* Faq section */
+/*  */
+
+.contact-faq {
+  width: 90%;
+  height: auto;
+  display: flex;
+  flex-direction: column;
+}
+/*  */
+/* Faq container */
+.contact-faq-container {
+  display: flex;
+  flex-direction: column;
+  gap: var(--space-1);
+  width: 100%;
+  min-height: 75vh;
+}
+/*  */
+/* Faq container item */
+.contact-faq-container-item {
+  display: flex;
+  background-color: var(--color-4);
+  padding: var(--space-3);
+  max-height: 200px;
+  overflow: hidden;
+  align-items: center;
+  border-radius: 25px;
+  transition:
+    background-color 0.5s var(--power-1-out),
+    max-height 0.5s var(--power-1-out);
+}
+
+.contact-faq-container-item h4 {
+  font-size: 1.5rem;
+  margin: 0;
+  color: var(--color-2);
+  margin-left: auto;
+  transition: font-size 0.5s var(--power-1-out);
+}
+
+.contact-faq-container-item-text {
+  width: 50%;
+  height: auto;
+}
+
+.contact-faq-container-item-text h3 {
+  font-size: 1.5rem;
+  margin: 0;
+  color: var(--color-1);
+  transition:
+    margin-bottom 0.5s var(--power-1-out),
+    color 0.5s var(--power-1-out);
+}
+
+.contact-faq-container-item-text p {
+  font-size: 1rem;
+  color: var(--font-middle);
+  max-height: 0;
+  overflow: hidden;
+  margin: 0;
+  transition: max-height 0.5s var(--power-1-out);
+}
+/* Hover state */
+.contact-faq-container-item.active {
+  background-color: var(--color-1);
+  max-height: 600px;
+}
+.contact-faq-container-item.active h4 {
+  font-size: 3rem;
+  color: var(--color-2);
+}
+
+.contact-faq-container-item-text.active h3 {
+  margin-bottom: var(--space-2);
+  color: var(--color-4);
+}
+
+.contact-faq-container-item-text.active p {
+  color: var(--color-3);
+  max-height: 300px;
+}
+
+@media (max-width: 1200px) {
+  .contact-faq-container-item-text {
+    width: 70%;
+  }
+}
