+++ Portfolio/src/views/contact/hero/heroAni.jsx
@@ -2,7 +2,7 @@ import { animateBlocks, animateText } from "@/utils/animations/animations";
 import { gsap } from "gsap/gsap-core";
 import { useEffect } from "react";
 
-function ContactAnimation() {
+function HeroAni() {
   useEffect(() => {
     const timeline = gsap.timeline();
     const header1 = document.querySelector("#test1");
@@ -44,20 +44,6 @@ function ContactAnimation() {
       timeline,
     );
   }, []);
-  useEffect(() => {
-    const faqCards = document.querySelectorAll(".contact-faq-container-item");
-    animateBlocks(
-      { start: 25, end: 0, type: "y" },
-      { el: "top", scroll: "90%" },
-      { el: "bottom", scroll: "40%" },
-      faqCards,
-      {
-        duration: 0.75,
-        easing: "power2.out",
-        stagger: 0.15,
-      },
-    );
-  }, []);
 }
 
-export default ContactAnimation;
+export default HeroAni;
