+++ Portfolio/src/views/skills/headerIntro/headerIntro.css
@@ -1,14 +1,3 @@
-/* The container for the entire page */
-.main-skill-container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 100%;
-  margin: 7.5rem 0;
-  gap: var(--space-7);
-}
 /*  */
 /* Top header container start */
 /*  */
