+++ Portfolio/src/views/skills/skill.css
@@ -9,320 +9,3 @@
   margin: 7.5rem 0;
   gap: var(--space-7);
 }
-/*  */
-/* Top header container start */
-/*  */
-.skill-header {
-  display: flex;
-  flex-direction: column;
-  width: 90%;
-  height: auto;
-  gap: var(--space-2);
-}
-/* Intro section */
-/*  */
-.skill-header-intro {
-  display: flex;
-  height: fit-content;
-  width: 100%;
-  padding: 0;
-  margin: 0;
-  align-items: flex-end;
-}
-/* Leftside text */
-.skill-header-intro-text {
-  width: auto;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  align-content: first baseline;
-  gap: auto;
-}
-
-.skill-header-intro-text h1 {
-  color: var(--color-2);
-  font-size: 4rem;
-  font-weight: 600;
-  margin: 0;
-  margin-bottom: var(--space-1);
-}
-
-.skill-header-intro-text h2 {
-  color: var(--font-middle);
-  font-size: 1.5rem;
-  font-weight: 500;
-  transform: translateX(15px);
-  margin: 0;
-}
-/* Rightside button */
-.skill-header-intro button {
-  border: none;
-  border-radius: 20px;
-  background-color: var(--color-1);
-  width: fit-content;
-  height: fit-content;
-  padding: var(--space-1) var(--space-3);
-  margin-left: auto;
-}
-.skill-header-intro h3 {
-  color: var(--color-3);
-  font-weight: 400;
-}
-/* Image area */
-/*  */
-
-.skill-header img {
-  width: 100%;
-  height: 25rem;
-  object-fit: cover;
-  display: block;
-  border-radius: 25px;
-}
-
-/* Stats */
-/*  */
-.skill-header-stats {
-  width: 100%;
-  display: flex;
-  height: auto;
-  flex-direction: row;
-  gap: var(--space-2);
-}
-
-.skill-header-stats-item {
-  width: 100%;
-  height: auto;
-  background-color: var(--color-4);
-  border-radius: 25px;
-  padding: var(--space-2) var(--space-3);
-}
-.skill-header-stats-item:nth-child(3) {
-  background-color: var(--color-1);
-}
-
-.skill-header-stats-item:nth-child(3) h3 {
-  color: var(--color-3);
-}
-
-.skill-header-stats-item:nth-child(3) h4 {
-  color: var(--color-5);
-}
-
-.skill-header-stats-item h3 {
-  color: var(--color-1);
-  font-weight: 600;
-  margin: 0;
-  margin-bottom: var(--space-1);
-  font-size: 2rem;
-}
-
-.skill-header-stats-item h4 {
-  font-weight: 500;
-  color: var(--font-middle);
-  margin: 0;
-  font-size: 0.85rem;
-}
-
-/*  */
-/* Past work container start */
-/*  */
-
-.skill-past {
-  width: 90%;
-  height: auto;
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-8);
-}
-
-.skill-past-item {
-  width: 100%;
-  height: auto;
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-1);
-}
-/*  */
-/* Main container large iamge small text contaienr */
-.skill-past-item-main {
-  display: flex;
-  overflow: hidden;
-  border-radius: 35px;
-}
-
-.skill-past-item-main.reverse {
-  flex-direction: row-reverse;
-}
-/* Left text part on default */
-.skill-past-item-main-text {
-  display: flex;
-  width: 35%;
-  flex-direction: column;
-  background-color: var(--color-4);
-  padding: var(--space-3) var(--space-2);
-}
-
-.skill-past-item-main-text h3 {
-  font-size: 2.5rem;
-  color: var(--color-2);
-  font-weight: 600;
-  margin: var(--space-1) 0;
-}
-
-.skill-past-item-main-text p {
-  font-size: 1.1rem;
-  color: var(--font-middle);
-}
-/* Right image part on default */
-.skill-past-item-main-image {
-  width: 65%;
-  display: flex;
-  flex-direction: row;
-  padding: var(--space-1);
-  position: relative;
-}
-
-.skill-past-item-main.reverse .skill-past-item-main-image {
-  flex-direction: row-reverse;
-}
-
-.skill-past-item-main-image img {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  width: 100%;
-  height: 100%;
-  transform: translate(-50%, -50%);
-  object-fit: cover;
-}
-
-.skill-past-item-main-image::before {
-  content: "";
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.25);
-
-  z-index: 10;
-}
-
-.skill-past-item-main-image-link {
-  z-index: 15;
-  width: 4rem;
-  height: 4rem;
-  background-color: var(--color-1);
-  border: none;
-  border-radius: 50%;
-  margin-top: auto;
-}
-
-.skill-past-item-main-image-link svg {
-  width: 2rem;
-  width: 2rem;
-  stroke: var(--color-3);
-  transform: rotate(225deg);
-}
-
-.skill-past-item-main-image-con {
-  height: auto;
-  display: flex;
-  flex-direction: column;
-  margin-left: auto;
-  margin-top: auto;
-  margin-bottom: auto;
-  gap: var(--space-1);
-  z-index: 15;
-}
-
-.skill-past-item-main.reverse .skill-past-item-main-image-con {
-  margin-left: 0;
-  margin-right: auto;
-}
-
-.skill-past-item-main-image-con span {
-  width: 2rem;
-  height: 2rem;
-  background-color: var(--color-4);
-  border: none;
-  border-radius: 50%;
-}
-
-.skill-past-item-main-image-con button {
-  width: 2rem;
-  height: 2rem;
-  background-color: var(--color-1);
-  border: none;
-  border-radius: 50%;
-}
-
-.skill-past-item-main-image-con button svg {
-  width: 1rem;
-  width: 1rem;
-  stroke: var(--color-3);
-}
-
-.skill-past-item-main-image-con button:nth-of-type(1) svg {
-  transform: rotate(180deg);
-}
-/*  */
-/* Stats section below main container */
-.skill-past-item-stats {
-  display: flex;
-  gap: var(--space-1);
-}
-
-.skill-past-item-stats-con {
-  display: flex;
-  flex-direction: column;
-  padding: var(--space-2);
-  background-color: var(--color-4);
-  border-radius: 25px;
-  width: 100%;
-}
-
-.skill-past-item-stats-con h4 {
-  font-size: 2rem;
-  color: var(--color-1);
-  font-weight: 500;
-  margin: var(--space-1) 0;
-}
-
-.skill-past-item-stats-con p {
-  font-size: 0.85rem;
-  color: var(--font-middle);
-  margin: 0;
-}
-/* Square stats area */
-.skill-past-item-stats-con-sq {
-  padding: 0;
-  display: flex;
-  width: 100%;
-  gap: var(--space-1);
-}
-
-.skill-past-item-stats-con-sq-it {
-  display: flex;
-  flex-direction: column;
-  flex: 1;
-  justify-content: center;
-  align-items: center;
-  border-radius: 20px;
-  height: 100%;
-  background-color: var(--color-1);
-}
-
-.skill-past-item-stats-con-sq-it h5 {
-  font-size: 1rem;
-  color: var(--color-3);
-  font-weight: 500;
-  margin: 0;
-}
-
-.skill-past-item-stats-con-sq-it h6 {
-  font-size: 0.75rem;
-  color: var(--color-5);
-  font-weight: 500;
-  margin: 0;
-}
